Fix saved UI parameter type sharing.
Move saved app parameter types into the shared package so the client route no longer imports a worker-only module, restoring worker typecheck after the bugfix pull.

Made-with: Cursor
